Enable the user to update the value of the timer The user can changes the value of the timer using the form provided For now, he needs to fill in the input with a value in milliseconds In the next iterations, we will use  - and + buttons to increase or decrease the timer
We created a class Timer to create / update new Timer easily
To avoid confusion the used formerly known as timer has been renamed
[delay, set_delay]

We added a new property to this state : new_timer. It's a boolean
and start with a value of false. We need this boolean to makes want a
new timer.
 - This property switch to "true" when the user complete the timer form
 - When this property is true :
    *we stop the current_timer (the one about to be triggered)
    *we create a new Timer with the updated value for the state delay
    *Thanks to side effects, the property start() of this new Timer is
    triggered after its inception.