We implement the function to display the solution The function right_fret() takes 2 inputs : - the note to play (A, B, C...G) - The number of the chord selected randomly by the app (6,5...1),
This function returns a number ranging from 1 to 12.
This number is the fret the should play to produce the note randomly
selected by the app, this is the solution.
The function has been coded but still need to be displayed on the UI
timely (using a timer probably or an useEffect -  tbd).