Implement and display the right solution The right_solution() is working well. Now we just need to set the timers properly. There are 3 timer: TIMER1 - The timer selected by the user TIMER2 - The timer before a new note is displayed TIMER3 - The timer to display the correct answer.
To make it works, we will need to add buffer time between each stage
EXAMPLE
TIMER1 : the user define a timer of 10s.
TIMER2 : the note is displayed during 15s
TIMER 3: The correct answer is displayed after 10s untill the next
note is displayed. In this example, the correct answer is then visible
for 5s.